+++ lens-labels/package.yaml
@@ -18,8 +18,9 @@ library:
   exposed-modules:
     - Lens.Labels
     - Lens.Labels.Unwrapped
-    - Lens.Prism
+    - Lens.Labels.Prism
   dependencies:
     - base >= 4.8 && < 4.11
     - ghc-prim >= 0.4 && < 0.6
     - profunctors >= 5.2
+    - tagged >= 0.8
