+++ proto-lens-protoc/package.yaml
@@ -62,7 +62,7 @@ library:
     - Lens.Family2 as Data.ProtoLens.Reexport.Lens.Family2,
     - Lens.Family2.Unchecked as Data.ProtoLens.Reexport.Lens.Family2.Unchecked,
     - Lens.Labels as Data.ProtoLens.Reexport.Lens.Labels,
-    - Lens.Prism as Data.ProtoLens.Reexport.Lens.Prism,
+    - Lens.Labels.Prism as Data.ProtoLens.Reexport.Lens.Labels.Prism,
     - Text.Read as Data.ProtoLens.Reexport.Text.Read
 
 executables:
