+++ proto-lens-tests/package.yaml
@@ -73,8 +73,6 @@ tests:
     dependencies:
       - proto-lens-tests
       - lens-labels
-      - tagged
-      - profunctors
     other-modules:
       - Proto.Oneof
       - Proto.Oneof'Fields
