+++ proto-lens-tests/tests/oneof_test.hs
@@ -5,31 +5,14 @@ import Proto.Oneof
 import Proto.Oneof'Fields
 import Data.ProtoLens
 import Lens.Family2 ((&), (.~), (^?), (%~), view)
-import Lens.Prism (_Just)
+import Lens.Labels.Prism (_Just, (#))
 import Test.Framework.Providers.HUnit (testCase)
 import Test.HUnit
-
--- Utilities for prism setting
-import Data.Functor.Identity (Identity (..))
-import Data.Profunctor.Unsafe ((#.), (.#))
-import Data.Tagged (Tagged (..))
 
 import Data.ProtoLens.TestUtil
 
 defFoo :: Foo
 defFoo = def
-
--- So we do not pull in the lens library we are putting some
--- utilities in here for setting prisms
-type Optic p f s t a b = p a (f b) -> p s (f t)
-
-type Optic' p f s a = Optic p f s s a a
-
-type AReview t b = Optic' Tagged Identity t b
-
-( # ) :: AReview t b -> b -> t
-( # ) p = runIdentity #. unTagged #. p .# Tagged .# Identity
-infixr 8 #
 
 main :: IO ()
 main = testMain
