+++ proto-lens/package.yaml
@@ -4,11 +4,11 @@ synopsis: A lens-based implementation of protocol buffers in Haskell.
 description: >
   The proto-lens library provides to protocol buffers using modern
   Haskell language and library patterns.  Specifically, it provides:
-
+   
   * Composable field accessors via lenses
-
+   
   * Simple field name resolution/overloading via type-level literals
-
+   
   * Type-safe reflection and encoding/decoding of messages via GADTs
 
 github: google/proto-lens/proto-lens
